+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Removed `days`, `hours`, `minutes`, `seconds`, `milliseconds`, `microseconds`, and `nanoseconds` methods (use `total_*` instead)
 - Removed `with_column` methods (use `with_columns` instead)
 - Removed `keep_name`, `prefix`, `suffix`, and `map_alias` methods (use `name` methods instead)
+- Removed `lstrip` and `rstrip` methods (use `strip_chars_start` and `strip_chars_end` instead)
 - Removed `cleared` methods (use `clear` instead)
 - Removed `count_match` methods (use `count_matches` instead)
 - Removed `avg` method (use `mean` instead)